requirements.txt
pandas==2.2.0
scipy==1.12.0
SQlAlchemy==2.0.25
python-dotenv==1.0.1
psycopg2==2.9.9
pymysql==1.1.0